requirements_base.txt
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
torch==1.9.1+cu111
torchvision==0.10.1+cu111
typing-extensions==3.10.0.2